Added Duality Game Engine
Duality is a 2D game engine that comes with a visual editor. Quick excerpt of some of its features:

- MIT Open Source
- Modular / Extensible, Plugin-Based
- Component-based game objects
- Live editing features
- Written entirely in C#
- The game core is a (non-shim) Portable Class Library, [Profile 111](http://danrigby.com/2014/05/14/supported-pcl-profiles-xamarin-for-visual-studio-2/). Non-portable backend code is loaded from plugins.

The project was started in the end of 2011 and is still in active development. The commercial game [Onikira](http://www.digitalfurnacegames.com/) is based on a fork of this project, a small community used it for several non-commercial games.

I am the maintainer of this project and am actively looking for long-term contributors.

**Official Info Page:** http://duality.adamslair.net/
**Source Code:** https://github.com/AdamsLair/duality
**Community Forum:** http://forum.adamslair.net/
**Blog:** http://blog.adamslair.net/