Add Mond scripting language
https://github.com/Rohansi/Mond

Mond is a scripting language that was written in pure C# to avoid deployment headaches other libraries have because they depend on native code. It maintained compatibility with Mono and now targets .NET Standard for portability.

It supports debugging the scripts using the existing remote debugger package, has a simple way to bind C# classes, a REPL with syntax highlighting, language features inspired by F#, and a website to try it out online.